refactor: extracting generation to another file
